Squashed 'src/vendor/clickhouse-cpp/' changes from cc538257..7deffc4c
7deffc4c change int128 lib

git-subtree-dir: src/vendor/clickhouse-cpp
git-subtree-split: 7deffc4c1b3b5e6861547ce34a8d80702915fc19

+++ clickhouse/columns/decimal.h
@@ -12,10 +12,10 @@ class ColumnDecimal : public Column {
 public:
     ColumnDecimal(size_t precision, size_t scale);
 
-    void Append(const Int128& value);
+    void Append(const BigInt& value);
     void Append(const std::string& value);
 
-    Int128 At(size_t i) const;
+    BigInt At(size_t i) const;
 
 public:
     void Append(ColumnRef column) override;
